+++ src/routes/graphql/queryType.ts
@@ -6,57 +6,40 @@ import { UserType } from './types/userType.js';
 import { UUIDType } from './types/uuid.js';
 import { getPost, getProfile, getMemberType } from './getsData.js';
 import { Context } from './types/types.js';
-import {
-  ResolveTree,
-  parseResolveInfo,
-  simplifyParsedResolveInfoFragmentWithType,
-} from 'graphql-parse-resolve-info';
+import { getUsers } from './resolvers/userResolver.js';
 
 export const QueryType = new GraphQLObjectType({
   name: 'Query',
   fields: {
     memberTypes: {
       type: new GraphQLList(MemberType),
-      resolve: (source, args, context: Context) => context.prisma.memberType.findMany(),
+      resolve: async (source, args, context: Context) => await context.prisma.memberType.findMany(),
     },
     memberType: {
       type: MemberType,
       args: {
         id: { type: new GraphQLNonNull(MemberTypeIdType) },
       },
-      resolve: (source, args: { id: string }, context) => getMemberType(args.id, context),
+      resolve: async (source, args: { id: string }, context) => await getMemberType(args.id, context),
     },
     posts: {
       type: new GraphQLList(PostType),
-      resolve: (source, args, context) => context.prisma.post.findMany(),
+      resolve: async (source, args, context) => await context.prisma.post.findMany(),
     },
     users: {
       type: new GraphQLList(UserType),
-      resolve: async (source, args, context, resolveInfo) => {
-        const parsedResolveInfoFragment = parseResolveInfo(resolveInfo) as ResolveTree;
-        const { fields }: { fields: { [key in string]: ResolveTree } } =
-          simplifyParsedResolveInfoFragmentWithType(parsedResolveInfoFragment, UserType);
-        const users = await context.prisma.user.findMany({
-          include: {
-            userSubscribedTo: !!fields.userSubscribedTo,
-            subscribedToUser: !!fields.subscribedToUser,
-          },
-        });
-
-        users.forEach((user) => context.userLoader.prime(user.id, user));
-        return users;
-      },
+      resolve: getUsers,
     },
     profiles: {
       type: new GraphQLList(ProfileType),
-      resolve: (source, args, context) => context.prisma.profile.findMany(),
+      resolve: async (source, args, context) => await context.prisma.profile.findMany(),
     },
     post: {
       type: PostType,
       args: {
         id: { type: new GraphQLNonNull(UUIDType) },
       },
-      resolve: async (source, args: { id: string }, context) => getPost(args.id, context),
+      resolve: async (source, args: { id: string }, context) => await getPost(args.id, context),
     },
     user: {
       type: UserType,
@@ -72,7 +55,7 @@ export const QueryType = new GraphQLObjectType({
         id: { type: new GraphQLNonNull(UUIDType) },
       },
       resolve: async (source, args: { id: string }, context) =>
-        getProfile(args.id, context),
+        await getProfile(args.id, context),
     },
   },
 });
